Use tabs over spaces again
Read and listened to some really good points
why tabs are superior to spaces.

Mostly for the visually impaired ones.

But the thing that struck me most is the following:

I keep arguing what the perfect indentation size is.
For me, it's 8 "spaces", but most people prefer 4 or 2.

If you use tabs, there is no point in arguing,
because everyone can setup the IDE as they wish and
veryone is happy.